pkgcheck.yml: Switch to hololeap/pkgcheck-action
Normally, the pkgcheck action silently ignores dependencies that are
unsatisfied in ::haskell but happened to exist in ::gentoo. This
modified pkgcheck-action prunes the ::gentoo tree of any packages that
exist in ::haskell to avoid these false negatives.

A side-effect of this is that it reveals very little coverage of
~arm64, ~ppc64, etc, though those architectures are keyworded on some
key packages.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>